Problem: Crash recovery for empty database not handled.
Solution: Handle the scenario where blocks collection is empty during crash recovery

+++ bigchaindb/commands/bigchaindb.py
@@ -159,7 +159,7 @@ def run_recover(b):
 
         # NOTE: the pre-commit state can only be ahead of the commited state
         # by 1 block
-        if latest_block['height'] < pre_commit['height']:
+        if latest_block and (latest_block['height'] < pre_commit['height']):
             query.delete_transactions(b.connection, pre_commit['transactions'])
